feat: enhance file processing and CLI options validation
- Introduce new CLI options parser for validation
- Implement path calculation service for output paths
- Replace batch processor service with file-batch-processor
- Update convertFilesUseCase to use new processing logic
- Remove deprecated services and streamline file handling
